do not rely on docker search, make remove more robust
git-svn-id: https://tor.ern.nps.edu/svn/proj/labtainer/trunk@800 ed162e55-3153-4343-a18d-e086fa355ed4

+++ scripts/labtainer-student/removelab.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 cont_list=$(docker ps -a | grep $1 | awk '{print $1}')
-print cont_list
-docker rm $cont_list
+if [[ ! -z "$cont_list" ]]; then
+    docker rm $cont_list
+fi
 image_list=$(docker images | grep $1 | awk '{print $1}')
-docker rmi -f $image_list
+if [[ ! -z "$image_list" ]]; then
+    docker rmi -f $image_list
+else
+    echo "No images for $1"
+fi
